requirements-integration.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
python-dotenv==0.12.0
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.16
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==1.0.1
zipp==0.6.0
